모든 스크립트 파일을 한데 모아, Scripts폴더를 정리한다
ChoiceSystem폴더에 Card와 Policy라는 이름의 폴더를 생성한다. ChoiceSystem.cs는 여기에 있다.
Card폴더에는 ChoiceCard.cs가 있고, Policy폴더에는 PolicySystem.cs와 PolicyIcon.cs가 있다.

GameFlow폴더에는 -Action으로 끝나는 이름의 스크립트들이 존재한다. 내용물이 하나임에도 굳이
폴더로 구분한 이유는 일관성을 부여하기 위함과, 각각의 경계를 명확하게 하기 위함이다.

General폴더에는 폴더로 구분할 수 없는 스크립트들이 모여있다. Train.cs와 Singleton.cs그리고
SoundSystem.cs가 이곳에 있다.
