+++ Assets/Scripts/ChoiceSystem/Card.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 82707fbd9369dc142b6345f5d2ab0874
+guid: 007d3cdfcdc4de949bfe2dcc9f4ba74c
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
