+++ Assets/Scripts/ChoiceSystem/Policy.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9286170b0b05ad2478ff369e5c5f72a0
+guid: a652f9497abbee347b2d0fb445483700
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
