+++ Assets/Scripts/ChoiceSystem/Policy/PolicyIcon.cs.meta
