+++ Assets/Scripts/General.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a9e85ceedd9eda4a89ada7647a42de2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
