+++ Assets/Scripts/General/Singleton.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a17aa641eca9d074497f459dbfa0614c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
