+++ Assets/Scripts/General/Singleton/Singleton.cs.meta
